+++ doc/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Highlighted Features
 
+* Display the current public body request email when notfing admins of a change
+  request (Gareth Rees)
 * Fix batch pages breaking due to empty `embargo_duration` value (Gareth Rees,
   Graeme Porteous)
 * Add quick links to browse latest requests and responses (Gareth Rees)